Update tg_fastRP.gsql - bug fix for print_results parameter
Bug fix for print_results parameter. When this parameter is set to TRUE, the script should print all node embeddings results. However all node embeddings will be zero due to the line s.@final_embedding_arr.clear(). The fix is to remove this line. Since this line is towards the end in the code, impact to memory consumption should be minimal.